Move data and tests out of the package
