+++ tests/common.py
@@ -4,6 +4,12 @@ __author__ = "Vanessa Sochat"
 __copyright__ = "Copyright 2016-2022, Vanessa Sochat"
 __license__ = "MIT"
 
+import os
+
+from deid.logger import bot
+
+from deid.utils import get_installdir
+
 
 def create_recipe(actions, fields=None, values=None):
     """
@@ -56,3 +62,28 @@ def get_file(dataset):
 
     dicom_files = get_files(dataset)
     return next(dicom_files)
+
+
+def get_dataset(dataset=None):
+    """
+    Get a dataset by name.
+
+    get_dataset will return some data provided by the application,
+    based on a user-provided label. In the future, we can add https endpoints
+    to retrieve online datasets.
+    """
+    data_base = os.path.join(os.path.dirname(__file__), "datasets")
+    valid_datasets = {
+        "dicom-cookies": os.path.join(data_base, "dicom-cookies"),
+        "animals": os.path.join(data_base, "animals"),
+        "humans": os.path.join(data_base, "humans"),
+        "ultrasounds": os.path.join(data_base, "ultrasounds"),
+    }
+
+    if dataset is not None:
+        # In case the user gave an extension
+        dataset = os.path.splitext(dataset)[0].lower()
+        if dataset in valid_datasets:
+            return valid_datasets[dataset]
+
+    bot.info("Valid datasets include: %s" % (",".join(list(valid_datasets.keys()))))
