+++ tests/datasets/dicom-cookies/image1.dcm
