+++ tests/datasets/dicom-cookies/image2.dcm
