+++ tests/datasets/dicom-cookies/image3.dcm
