+++ tests/datasets/dicom-cookies/image4.dcm
