+++ tests/datasets/dicom-cookies/image5.dcm
