+++ tests/datasets/dicom-cookies/image6.dcm
