+++ tests/datasets/dicom-cookies/image7.dcm
