+++ tests/datasets/ultrasounds/GREYSCALE_CINE.zip
