+++ tests/datasets/ultrasounds/GREYSCALE_IMAGE.dcm
