+++ tests/datasets/ultrasounds/RGB_IMAGE.dcm
