+++ tests/resources/filter_multiple_first_filter_match.dicom
