+++ tests/resources/filter_multiple_rule_innerop_true.dicom
