+++ tests/resources/filter_multiple_second_filter_match.dicom
