+++ tests/resources/filter_multiple_zero_filter_match.dicom
