+++ tests/resources/filter_single_rule_false.dicom
