+++ tests/resources/filter_single_rule_innerop_true.dicom
