+++ tests/resources/filter_single_rule_true.dicom
