+++ tests/resources/keepcoordinates.dicom
