+++ tests/resources/keepcoordinates_from.dicom
