+++ tests/resources/keepcoordinates_noaction.dicom
