+++ tests/resources/remove_all.dicom
