+++ tests/resources/remove_coordinates.dicom
