+++ tests/resources/remove_coordinates_multiple.dicom
