+++ tests/resources/remove_coordinates_multiple_filters.dicom
