+++ tests/resources/remove_coordinates_us.dicom
