+++ tests/resources/remove_coordinates_us_all.dicom
