+++ tests/test_clean.py
@@ -4,26 +4,33 @@ __author__ = "Vanessa Sochat"
 __copyright__ = "Copyright 2016-2022, Vanessa Sochat"
 __license__ = "MIT"
 
+
 """
-Test DICOM Cleaner - Images with varying pixel dimensions
+Test DICOM Cleaner
 """
 
 import os
 import shutil
 import tempfile
 import unittest
+from copy import deepcopy
 
+import pydicom
 from pydicom import read_file
 
-from deid.data import get_dataset
+from deid.config import DeidRecipe
+from deid.dicom.pixels import clean_pixel_data, has_burned_pixels
+from tests.common import get_file, get_dataset
 from deid.utils import get_installdir
 
+global generate_uid
 
-class TestCleanPizelDimensions(unittest.TestCase):
+
+class TestClean(unittest.TestCase):
     def setUp(self):
         self.pwd = get_installdir()
-        self.deidpath = os.path.abspath("%s/tests/resources/" % self.pwd)
-        self.dataset = get_dataset("ultrasounds")
+        self.deidpath = os.path.abspath("%s/../tests/resources/" % self.pwd)
+        self.dataset = get_dataset("animals")
         self.tmpdir = tempfile.mkdtemp()
         print("\n######################START######################")
 
@@ -31,15 +38,12 @@ class TestCleanPizelDimensions(unittest.TestCase):
         shutil.rmtree(self.tmpdir)
         print("\n######################END########################")
 
-    def test_4d_RGB_cine_clip(self):
-        """
-        Test the pixel cleaner to ensure pixels are appropriately deidentified
-        on "4D" images - RGB cine clips.  Pixel data will have the shape (frames, X, Y, channel)
-        """
+    def test_pixel_cleaner_remove_coordinates(self):
+        """Test the pixel cleaner to ensure it appropriately clears specified pixels."""
         from deid.dicom import DicomCleaner
 
-        dicom_file = get_file(self.dataset, "RGB_CINE.zip", self.tmpdir)
-        deid = os.path.join(self.deidpath, "remove_coordinates_us.dicom")
+        dicom_file = get_file(self.dataset)
+        deid = os.path.join(self.deidpath, "remove_coordinates.dicom")
 
         client = DicomCleaner(output_folder=self.tmpdir, deid=deid)
         out = client.detect(dicom_file)
@@ -56,193 +60,36 @@ class TestCleanPizelDimensions(unittest.TestCase):
         compare = inputpixels == outputpixels
         self.assertFalse(compare.all())
 
-        inputpixels[:, 0:500, 0:500, :] = 0
+        inputpixels[0:1024, 0:1024] = 0
         compare = inputpixels == outputpixels
         self.assertTrue(compare.all())
 
-    def test_3d_Greyscale_cine_clip(self):
-        """
-        Test the pixel cleaner to ensure pixels are appropriately deidentified
-        on "3D" images - greyscale cine clips.  Pixel data will have the shape (frames, X, Y)
-        """
-        from deid.dicom import DicomCleaner
+    def test_pixel_cleaner_remove_coordinates_dicom_file(self):
+        """Test the pixel cleaner to ensure it appropriately clears specified pixels."""
+        dicom_file_data = pydicom.read_file(get_file(self.dataset))
+        inputpixels = deepcopy(dicom_file_data.pixel_array)
 
-        dicom_file = get_file(self.dataset, "GREYSCALE_CINE.zip", self.tmpdir)
-        deid = os.path.join(self.deidpath, "remove_coordinates_us.dicom")
+        deid_path = os.path.join(self.deidpath, "remove_coordinates.dicom")
+        deid = DeidRecipe(deid_path)
 
-        client = DicomCleaner(output_folder=self.tmpdir, deid=deid)
-        out = client.detect(dicom_file)
+        out = has_burned_pixels(dicom_file_data, deid=deid)
         self.assertTrue(out["flagged"])
 
-        client.clean()
-        cleanedfile = client.save_dicom()
+        outputpixels = clean_pixel_data(dicom_file=dicom_file_data, results=out)
 
-        outputfile = read_file(cleanedfile)
-        outputpixels = outputfile.pixel_array
-
-        inputfile = read_file(dicom_file)
-        inputpixels = inputfile.pixel_array
         compare = inputpixels == outputpixels
         self.assertFalse(compare.all())
 
-        inputpixels[:, 0:500, 0:500] = 0
+        inputpixels[0:1024, 0:1024] = 0
         compare = inputpixels == outputpixels
         self.assertTrue(compare.all())
 
-    def test_3d_RGB_image(self):
-        """
-        Test the pixel cleaner to ensure pixels are appropriately deidentified
-        on "3D" images - RGB images.  Pixel data will have the shape (X, Y, channel)
-        """
+    def test_pixel_cleaner_remove_all(self):
+        """Test the pixel cleaner to ensure it appropriately clears all pixels."""
         from deid.dicom import DicomCleaner
 
-        dicom_file = get_file(self.dataset, "RGB_IMAGE.dcm")
-        deid = os.path.join(self.deidpath, "remove_coordinates_us.dicom")
-
-        client = DicomCleaner(output_folder=self.tmpdir, deid=deid)
-        out = client.detect(dicom_file)
-        self.assertTrue(out["flagged"])
-
-        client.clean()
-        cleanedfile = client.save_dicom()
-
-        outputfile = read_file(cleanedfile)
-        outputpixels = outputfile.pixel_array
-
-        inputfile = read_file(dicom_file)
-        inputpixels = inputfile.pixel_array
-        compare = inputpixels == outputpixels
-        self.assertFalse(compare.all())
-
-        inputpixels[0:500, 0:500, :] = 0
-        compare = inputpixels == outputpixels
-        self.assertTrue(compare.all())
-
-    def test_2d_Greyscale_image(self):
-        """
-        Test the pixel cleaner to ensure pixels are appropriately deidentified
-        on "2D" images - Greyscale images.  Pixel data will have the shape (X, Y)
-        """
-        from deid.dicom import DicomCleaner
-
-        dicom_file = get_file(self.dataset, "GREYSCALE_IMAGE.dcm")
-        deid = os.path.join(self.deidpath, "remove_coordinates_us.dicom")
-
-        client = DicomCleaner(output_folder=self.tmpdir, deid=deid)
-        out = client.detect(dicom_file)
-        self.assertTrue(out["flagged"])
-
-        client.clean()
-        cleanedfile = client.save_dicom()
-
-        outputfile = read_file(cleanedfile)
-        outputpixels = outputfile.pixel_array
-
-        inputfile = read_file(dicom_file)
-        inputpixels = inputfile.pixel_array
-        compare = inputpixels == outputpixels
-        self.assertFalse(compare.all())
-
-        inputpixels[0:500, 0:500] = 0
-        compare = inputpixels == outputpixels
-        self.assertTrue(compare.all())
-
-    def test_4d_RGB_cine_clip_all(self):
-        """
-        Test the pixel cleaner to ensure pixels are appropriately deidentified - all keyword -
-        on "4D" images - RGB cine clips.  Pixel data will have the shape (frames, X, Y, channel)
-        """
-        from deid.dicom import DicomCleaner
-
-        dicom_file = get_file(self.dataset, "RGB_CINE.zip", self.tmpdir)
-        deid = os.path.join(self.deidpath, "remove_coordinates_us_all.dicom")
-
-        client = DicomCleaner(output_folder=self.tmpdir, deid=deid)
-        out = client.detect(dicom_file)
-        self.assertTrue(out["flagged"])
-
-        client.clean()
-        cleanedfile = client.save_dicom()
-
-        outputfile = read_file(cleanedfile)
-        outputpixels = outputfile.pixel_array
-
-        inputfile = read_file(dicom_file)
-        inputpixels = inputfile.pixel_array
-        compare = inputpixels == outputpixels
-        self.assertFalse(compare.all())
-
-        inputpixels[:, :, :, :] = 0
-        compare = inputpixels == outputpixels
-        self.assertTrue(compare.all())
-
-    def test_3d_Greyscale_cine_clip_all(self):
-        """
-        Test the pixel cleaner to ensure pixels are appropriately deidentified - all keyword -
-        on "3D" images - greyscale cine clips.  Pixel data will have the shape (frames, X, Y)
-        """
-        from deid.dicom import DicomCleaner
-
-        dicom_file = get_file(self.dataset, "GREYSCALE_CINE.zip", self.tmpdir)
-        deid = os.path.join(self.deidpath, "remove_coordinates_us_all.dicom")
-
-        client = DicomCleaner(output_folder=self.tmpdir, deid=deid)
-        out = client.detect(dicom_file)
-        self.assertTrue(out["flagged"])
-
-        client.clean()
-        cleanedfile = client.save_dicom()
-
-        outputfile = read_file(cleanedfile)
-        outputpixels = outputfile.pixel_array
-
-        inputfile = read_file(dicom_file)
-        inputpixels = inputfile.pixel_array
-        compare = inputpixels == outputpixels
-        self.assertFalse(compare.all())
-
-        inputpixels[:, :, :] = 0
-        compare = inputpixels == outputpixels
-        self.assertTrue(compare.all())
-
-    def test_3d_RGB_image_all(self):
-        """
-        Test the pixel cleaner to ensure pixels are appropriately deidentified - all keyword -
-        on "3D" images - RGB images.  Pixel data will have the shape (X, Y, channel)
-        """
-        from deid.dicom import DicomCleaner
-
-        dicom_file = get_file(self.dataset, "RGB_IMAGE.dcm")
-        deid = os.path.join(self.deidpath, "remove_coordinates_us_all.dicom")
-
-        client = DicomCleaner(output_folder=self.tmpdir, deid=deid)
-        out = client.detect(dicom_file)
-        self.assertTrue(out["flagged"])
-
-        client.clean()
-        cleanedfile = client.save_dicom()
-
-        outputfile = read_file(cleanedfile)
-        outputpixels = outputfile.pixel_array
-
-        inputfile = read_file(dicom_file)
-        inputpixels = inputfile.pixel_array
-        compare = inputpixels == outputpixels
-        self.assertFalse(compare.all())
-
-        inputpixels[:, :, :] = 0
-        compare = inputpixels == outputpixels
-        self.assertTrue(compare.all())
-
-    def test_2d_Greyscale_image_all(self):
-        """
-        Test the pixel cleaner to ensure pixels are appropriately deidentified - all keyword -
-        on "2D" images - Greyscale images.  Pixel data will have the shape (X, Y)
-        """
-        from deid.dicom import DicomCleaner
-
-        dicom_file = get_file(self.dataset, "GREYSCALE_IMAGE.dcm")
-        deid = os.path.join(self.deidpath, "remove_coordinates_us_all.dicom")
+        dicom_file = get_file(self.dataset)
+        deid = os.path.join(self.deidpath, "remove_all.dicom")
 
         client = DicomCleaner(output_folder=self.tmpdir, deid=deid)
         out = client.detect(dicom_file)
@@ -263,15 +110,134 @@ class TestCleanPizelDimensions(unittest.TestCase):
         compare = inputpixels == outputpixels
         self.assertTrue(compare.all())
 
+    def test_pixel_cleaner_keepcoordinates_noaction(self):
+        """Test the pixel cleaner to ensure that a keepcoordinates with no removecoordinates has no impact on the pixels."""
+        from deid.dicom import DicomCleaner
 
-def get_file(dataset, image, tempdir=None):
-    """
-    Helper to get a dicom file
-    """
-    from deid.dicom import get_files
+        dicom_file = get_file(self.dataset)
+        deid = os.path.join(self.deidpath, "keepcoordinates_noaction.dicom")
 
-    dicom_files = get_files(dataset, pattern=image, tempdir=tempdir)
-    return next(dicom_files)
+        client = DicomCleaner(output_folder=self.tmpdir, deid=deid)
+        out = client.detect(dicom_file)
+        self.assertTrue(out["flagged"])
+
+        client.clean()
+        cleanedfile = client.save_dicom()
+
+        outputfile = read_file(cleanedfile)
+        outputpixels = outputfile.pixel_array
+
+        inputfile = read_file(dicom_file)
+        inputpixels = inputfile.pixel_array
+        compare = inputpixels == outputpixels
+        self.assertTrue(compare.all())
+
+    def test_pixel_cleaner_keepcoordinates(self):
+        """Test the pixel cleaner to ensure that a keepcoordinates retains appropriate pixels."""
+        from deid.dicom import DicomCleaner
+
+        dicom_file = get_file(self.dataset)
+        deid = os.path.join(self.deidpath, "keepcoordinates.dicom")
+
+        client = DicomCleaner(output_folder=self.tmpdir, deid=deid)
+        out = client.detect(dicom_file)
+        self.assertTrue(out["flagged"])
+
+        client.clean()
+        cleanedfile = client.save_dicom()
+
+        outputfile = read_file(cleanedfile)
+        outputpixels = outputfile.pixel_array
+
+        inputfile = read_file(dicom_file)
+        inputpixels = inputfile.pixel_array
+        compare = inputpixels == outputpixels
+        self.assertFalse(compare.all())
+
+        compare = inputpixels[0:1024, 0:1024] == outputpixels[0:1024, 0:1024]
+        self.assertTrue(compare.all())
+
+    def test_pixel_cleaner_remove_multiple(self):
+        """Test the pixel cleaner to ensure that multiple remove coordinates in the same filter remove the appropriate pixels."""
+        from deid.dicom import DicomCleaner
+
+        dicom_file = get_file(self.dataset)
+        deid = os.path.join(self.deidpath, "remove_coordinates_multiple.dicom")
+
+        client = DicomCleaner(output_folder=self.tmpdir, deid=deid)
+        out = client.detect(dicom_file)
+        self.assertTrue(out["flagged"])
+
+        client.clean()
+        cleanedfile = client.save_dicom()
+
+        outputfile = read_file(cleanedfile)
+        outputpixels = outputfile.pixel_array
+
+        inputfile = read_file(dicom_file)
+        inputpixels = inputfile.pixel_array
+        compare = inputpixels == outputpixels
+        self.assertFalse(compare.all())
+
+        inputpixels[0:10, 0:10] = 0
+        inputpixels[10:20, 10:20] = 0
+        compare = inputpixels == outputpixels
+        self.assertTrue(compare.all())
+
+    def test_pixel_cleaner_remove_multiple_filters(self):
+        """Test the pixel cleaner to ensure that multiple remove coordinates in different filters remove the appropriate pixels."""
+        from deid.dicom import DicomCleaner
+
+        dicom_file = get_file(self.dataset)
+        deid = os.path.join(self.deidpath, "remove_coordinates_multiple_filters.dicom")
+
+        client = DicomCleaner(output_folder=self.tmpdir, deid=deid)
+        out = client.detect(dicom_file)
+        self.assertTrue(out["flagged"])
+
+        client.clean()
+        cleanedfile = client.save_dicom()
+
+        outputfile = read_file(cleanedfile)
+        outputpixels = outputfile.pixel_array
+
+        inputfile = read_file(dicom_file)
+        inputpixels = inputfile.pixel_array
+        compare = inputpixels == outputpixels
+        self.assertFalse(compare.all())
+
+        inputpixels[0:10, 0:10] = 0
+        inputpixels[10:20, 10:20] = 0
+        compare = inputpixels == outputpixels
+        self.assertTrue(compare.all())
+
+    def test_pixel_cleaner_keepcoordinates_from(self):
+        """Test the pixel cleaner to ensure that multiple keep coordinates retrieved from a dicom field are appropriately retained."""
+        from deid.dicom import DicomCleaner
+
+        dicom_file = get_file(self.dataset)
+        deid = os.path.join(self.deidpath, "keepcoordinates_from.dicom")
+
+        client = DicomCleaner(output_folder=self.tmpdir, deid=deid)
+        out = client.detect(dicom_file)
+        self.assertTrue(out["flagged"])
+
+        client.clean()
+        cleanedfile = client.save_dicom()
+
+        outputfile = read_file(cleanedfile)
+        outputpixels = outputfile.pixel_array
+
+        inputfile = read_file(dicom_file)
+        inputpixels = inputfile.pixel_array
+        compare = inputpixels == outputpixels
+        self.assertFalse(compare.all())
+
+        inputpixels[1000:2000, 0:1000] = 0
+        inputpixels[0:1000, 1000:2000] = 0
+
+        compare = inputpixels[0:2000, 0:2000] == outputpixels[0:2000, 0:2000]
+        self.assertTrue(compare.all())
 
 
 if __name__ == "__main__":
