+++ tests/test_dicom_fields.py
@@ -13,9 +13,8 @@ import shutil
 import tempfile
 import unittest
 
-from deid.data import get_dataset
 from deid.dicom.fields import get_fields
-from deid.tests.common import get_dicom
+from tests.common import get_dicom, get_dataset
 from deid.utils import get_installdir
