+++ tests/test_dicom_utils.py
@@ -13,8 +13,7 @@ import shutil
 import tempfile
 import unittest
 
-from deid.data import get_dataset
-from deid.tests.common import get_dicom
+from tests.common import get_dicom, get_dataset
 from deid.utils import get_installdir
 
 global generate_uid
