+++ tests/test_replace_identifiers.py
@@ -17,10 +17,9 @@ from collections import OrderedDict
 from pydicom import read_file
 from pydicom.sequence import Sequence
 
-from deid.data import get_dataset
 from deid.dicom import get_identifiers, replace_identifiers
 from deid.dicom.parser import DicomParser
-from deid.tests.common import create_recipe, get_file
+from tests.common import create_recipe, get_file, get_dataset
 from deid.utils import get_installdir
 
 global generate_uid
